Trying to work hit detection
I have set hit detection for all of the circles of the first question.
Nothing happens when they hit yet. I want to get it so that when my
mouse hits the wrong answers it says “NO” and when i hit the correct
answer it goes onto the next question. The only problem is that i don’t
know how to change the location of the question after each question. I
also do not know how to switch the screen after the correct answer is
had.